requirements.txt
# Core dependencies
aiohttp>=3.8.0
requests>=2.25.0
pyyaml>=6.0

# For data handling
python-dateutil>=2.8.0

# Optional: For enhanced reporting
jinja2>=3.0.0  # For HTML templates
pandas>=1.3.0  # For data analysis
matplotlib>=3.3.0  # For charts in reports

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950